refactor(docs): popover dx
