+++ apps/docs/content/components/popover/backdrop.ts
@@ -1,48 +1,4 @@
-const App = `import {Popover, PopoverTrigger, PopoverContent, Button, Input} from "@nextui-org/react";
-
-export default function App() {
-
-  const backdrops = ["opaque", "blur", "transparent"];
-
-  const content = (
-    <PopoverContent className="w-[240px]">
-      {(titleProps) => (
-        <div className="px-1 py-2 w-full">
-          <p className="text-small font-bold text-foreground" {...titleProps}>
-            Dimensions
-          </p>
-          <div className="mt-2 flex flex-col gap-2 w-full">
-            <Input defaultValue="100%" label="Width" size="sm" variant="bordered" />
-            <Input defaultValue="300px" label="Max. width" size="sm" variant="bordered" />
-            <Input defaultValue="24px" label="Height" size="sm" variant="bordered" />
-            <Input defaultValue="30px" label="Max. height" size="sm" variant="bordered" />
-          </div>
-        </div>
-      )}
-    </PopoverContent>
-  )
-
-  return (
-    <div className="flex flex-wrap gap-4">
-      {backdrops.map((backdrop) => (
-        <Popover
-          key={backdrop}
-          showArrow
-          offset={10}
-          placement="bottom"
-          backdrop={backdrop}
-        >
-          <PopoverTrigger>
-            <Button color="warning" variant="flat" className="capitalize">
-              {backdrop}
-            </Button>
-          </PopoverTrigger>
-          {content}
-        </Popover>
-      ))}
-    </div>
-  );
-}`;
+import App from "./backdrop.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
