+++ apps/docs/content/components/popover/custom-trigger.ts
@@ -1,78 +1,7 @@
-const UserTwitterCard = `import { Avatar, Button, Card, CardBody, CardFooter, CardHeader } from "@nextui-org/react";
-
-export const UserTwitterCard = () => {
-  const [isFollowed, setIsFollowed] = React.useState(false);
-
-  return (
-    <Card shadow="none" className="max-w-[300px] border-none bg-transparent">
-      <CardHeader className="justify-between">
-        <div className="flex gap-3">
-          <Avatar isBordered radius="full" size="md" src="https://i.pravatar.cc/150?u=a04258114e29026702d" />
-          <div className="flex flex-col items-start justify-center">
-            <h4 className="text-small font-semibold leading-none text-default-600">Zoey Lang</h4>
-            <h5 className="text-small tracking-tight text-default-500">@zoeylang</h5>
-          </div>
-        </div>
-        <Button
-          className={isFollowed ? "bg-transparent text-foreground border-default-200" : ""}
-          color="primary"
-          radius="full"
-          size="sm"
-          variant={isFollowed ? "bordered" : "solid"}
-          onPress={() => setIsFollowed(!isFollowed)}
-        >
-          {isFollowed ? "Unfollow" : "Follow"}
-        </Button>
-      </CardHeader>
-      <CardBody className="px-3 py-0">
-        <p className="text-small pl-px text-default-500">
-          Full-stack developer, @getnextui lover she/her
-          <span aria-label="confetti" role="img">
-            🎉
-          </span>
-        </p>
-      </CardBody>
-      <CardFooter className="gap-3">
-        <div className="flex gap-1">
-          <p className="font-semibold text-default-600 text-small">4</p>
-          <p className=" text-default-500 text-small">Following</p>
-        </div>
-        <div className="flex gap-1">
-          <p className="font-semibold text-default-600 text-small">97.1K</p>
-          <p className="text-default-500 text-small">Followers</p>
-        </div>
-      </CardFooter>
-    </Card>
-  );
-};`;
-
-const App = `import {Popover, PopoverTrigger, PopoverContent, Button, User} from "@nextui-org/react";
-import {UserTwitterCard} from "./UserTwitterCard";
-
-export default function App() {
-  return (
-    <Popover showArrow placement="bottom">
-      <PopoverTrigger>
-        <User   
-          as="button"
-          name="Zoe Lang"
-          description="Product Designer"
-          className="transition-transform"
-          avatarProps={{
-            src: "https://i.pravatar.cc/150?u=a04258114e29026702d"
-          }}
-        />
-      </PopoverTrigger>
-      <PopoverContent className="p-1">
-        <UserTwitterCard />
-      </PopoverContent>
-    </Popover>
-  );
-}`;
+import App from "./custom-trigger.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
-  "/UserTwitterCard.jsx": UserTwitterCard,
 };
 
 export default {
