+++ apps/docs/content/components/popover/placements.ts
@@ -1,45 +1,4 @@
-const App = `import {Popover, PopoverTrigger, PopoverContent, Button} from "@nextui-org/react";
-
-export default function App() {
-  const content = (
-    <PopoverContent>
-      <div className="px-1 py-2">
-        <div className="text-small font-bold">Popover Content</div>
-        <div className="text-tiny">This is the popover content</div>
-      </div>
-    </PopoverContent>
-  );
-
-  const placements = [
-    "top-start",
-    "top",
-    "top-end",
-    "bottom-start",
-    "bottom",
-    "bottom-end",
-    "right-start",
-    "right",
-    "right-end",
-    "left-start",
-    "left",
-    "left-end",
-  ];
-
-  return (
-    <div className="flex flex-wrap md:inline-grid md:grid-cols-3 gap-4">
-      {placements.map((placement) => (
-        <Popover key={placement} placement={placement} color="secondary">
-          <PopoverTrigger>
-            <Button color="secondary" variant="flat" className="capitalize">
-              {placement.replace("-", " ")}
-            </Button>
-          </PopoverTrigger>
-          {content}
-        </Popover>
-      ))}
-    </div>
-  );
-}`;
+import App from "./placements.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
